+++ examples/relations.py
@@ -8,10 +8,8 @@ to get this related objects
 import asyncio
 
 from tortoise import Tortoise, fields
-from tortoise.backends.sqlite.client import SqliteClient
 from tortoise.exceptions import NoValuesFetched
 from tortoise.models import Model
-from tortoise.utils import generate_schema
 
 
 class Tournament(Model):
@@ -43,10 +41,8 @@ class Team(Model):
 
 
 async def run():
-    client = SqliteClient('example_relations.sqlite3')
-    await client.create_connection()
-    Tortoise.init(client)
-    await generate_schema(client)
+    await Tortoise.init(config_file='config.json')
+    await Tortoise.generate_schemas()
 
     tournament = Tournament(name='New Tournament')
     await tournament.save()