+++ tortoise/backends/asyncpg/__init__.py
@@ -1,0 +1,3 @@
+from .client import AsyncpgDBClient
+
+client_class = AsyncpgDBClient