+++ tortoise/backends/sqlite/__init__.py
@@ -1,0 +1,3 @@
+from .client import SqliteClient
+
+client_class = SqliteClient